# wiki-stress/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ukrainian Stress Lexicon Lookup API
  description: >
    Read-only HTTP lookup service over the published Ukrainian stress
    lexicon. This service performs no dump parsing, morphology extraction,
    import, or migration — see openspec/openspec/project.md for the
    Python/Go architecture boundary.
  version: "1.0.0"
servers:
  - url: /
paths:
...
  /v1/lemmas/{lemma}/forms:
    get:
      summary: Paginated exact retrieval of stored forms for a lemma
      operationId: lemmaForms
      parameters:
        - name: lemma
          in: path
          required: true
          schema:
            type: string
            minLength: 1
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            description: Clamped server-side to the configured page-size maximum.
      responses:
        "200":
          description: Paginated forms for the lemma.
          content:
            application/json:
              schema:
                type: object
                required: [lemma, forms, offset, limit]
                properties:
                  lemma:
                    type: string
                  forms:
                    type: array
                    items:
                      $ref: "#/components/schemas/LemmaForm"
                  offset:
                    type: integer
                  limit:
                    type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  schemas:
...
    LemmaForm:
      type: object
      required: [form, stressed_form, grammatical_tags, confidence]
      properties:
        form:
          type: string
        stressed_form:
          type: string
        grammatical_tags:
          type: array
          items:
            type: string
        confidence:
          type: number
          format: float
...
    Error:
      type: object
      required: [code, message, request_id]
      description: >
        Stable error contract. Never contains SQL text, credentials, or a
        stack trace.
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServiceUnavailable:
      description: >
        PostgreSQL is unreachable or no dataset is currently published.
        Readiness fails in this state; liveness does not.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"